backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.20

openai>=1.6.1,<2.0.0
httpx==0.25.2
python-multipart==0.0.6
youtube-search-python==1.6.6
beautifulsoup4==4.12.2
requests==2.31.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2